fix: credit type change integer to long

+++ src/main/java/com/bitharmony/comma/donation/dto/DonationOnceRequestDto.java
@@ -12,7 +12,7 @@ public record DonationOnceRequestDto(
         String artistName,
 
         @NotEmpty
-        Integer amount,
+        Long amount,
 
         String message,
